perf: Greatly reduce memory usage by loading only the session currently viewed (#34)
When there is a lot of sessions with many tabs (think of 400+ sessions with each 150 tab+), memory usage increases greatly. There is no reason to load all the data for every session loaded, instead just load the session that is being viewed. This reduces memory usage by a very large margin.